quell warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@17510 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/other/curl/lib/tftp.c
@@ -214,14 +214,14 @@ static CURLcode tftp_set_timeouts(tftp_state_data_t *state)
     timeout = maxtime ;
 
     /* Average restart after 5 seconds */
-    state->retry_max = timeout/5;
+    state->retry_max = (int)(timeout/5);
 
     if(state->retry_max < 1)
       /* avoid division by zero below */
       state->retry_max = 1;
 
     /* Compute the re-start interval to suit the timeout */
-    state->retry_time = timeout/state->retry_max;
+    state->retry_time = (int)(timeout/state->retry_max);
     if(state->retry_time<1)
       state->retry_time=1;
 
@@ -238,7 +238,7 @@ static CURLcode tftp_set_timeouts(tftp_state_data_t *state)
     timeout = maxtime/10 ;
 
     /* Average reposting an ACK after 15 seconds */
-    state->retry_max = timeout/15;
+    state->retry_max = (int)(timeout/15);
   }
   /* But bound the total number  */
   if(state->retry_max<3)
@@ -248,7 +248,7 @@ static CURLcode tftp_set_timeouts(tftp_state_data_t *state)
     state->retry_max=50;
 
   /* Compute the re-ACK interval to suit the timeout */
-  state->retry_time = timeout/state->retry_max;
+  state->retry_time = (int)(timeout/state->retry_max);
   if(state->retry_time<1)
     state->retry_time=1;
